requirements.txt
streamlit==1.36.0
biopython==1.83
requests>=2.31
torch>=2.1
numpy>=1.26
